Refactor of relative address int>double (for offset grids)

+++ mcomp_Java/src/common/datatypes/map/griddedMap/RelativeAddress.java
@@ -1,20 +1,24 @@
-/**
- * 
- */
 package common.datatypes.map.griddedMap;
 
 /**
+ * Tuple holding the offset (x/y) for each address Carier Object
+ * 
  * @author David Avery 15823926
+ * @version 1.0
+ * @since 2018-04-07
+ * 
+ * @see GridDesign
  *
  */
 class RelativeAddress {
-  public final int neighbourXOffset;
-  public final int neighbourYOffset;
+  public final double neighbourXOffset;
+  public final double neighbourYOffset;
 
   /**
+   * Constructor
    * 
    */
-  public RelativeAddress(int x, int y) {
+  public RelativeAddress(double x, double y) {
     neighbourXOffset = x;
     neighbourYOffset = y;
   }
